[mod] add searx/webutils.py
contains utility functions and classes used only by webapp.py
